Hookup more of the new media page
