+++ app/controllers/media/show.js
@@ -3,8 +3,6 @@ import computed, { alias } from 'ember-computed';
 import get from 'ember-metal/get';
 import service from 'ember-service/inject';
 import { image } from 'client/helpers/image';
-import moment from 'moment';
-/* global humanizeDuration */
 
 export default Controller.extend({
   media: alias('model'),
@@ -23,22 +21,9 @@ export default Controller.extend({
     }
   }).readOnly(),
 
-  airedLongerThanOneDay: computed('media.startDate', 'media.endDate', {
-    get() {
-      return !(get(this, 'media.startDate').isSame(get(this, 'media.endDate')));
-    }
-  }).readOnly(),
-
   totalProgressText: computed('media.unitCount', {
     get() {
       return get(this, 'media.unitCount') || '-';
-    }
-  }).readOnly(),
-
-  totalTime: computed('media.episodeLength', {
-    get() {
-      const time = moment.duration(get(this, 'media.episodeCount') * get(this, 'media.episodeLength'), 'minutes');
-      return humanizeDuration(time.asMilliseconds(), { largest: 1 });
     }
   }).readOnly(),
